+++ Lib/eacp/Graphics/Tray/TrayIcon-Windows.cpp
@@ -3,13 +3,12 @@
 #include "TrayIcon.h"
 #include "../Helpers/StringUtils-Windows.h"
 #include "../Helpers/DarkMode-Windows.h"
+#include "../Helpers/ImageConversion-Windows.h"
 #include <eacp/Core/App/AppEnvironment.h>
 #include <eacp/Core/Threads/EventLoop.h>
-#include <eacp/SIMD/SIMD.h>
 
 #include <shellapi.h>
 
-#include <cstdint>
 #include <unordered_map>
 
 namespace eacp::Graphics
@@ -21,69 +20,6 @@ static bool trayWindowClassRegistered = false;
 // The notification-area callback message and the icon's per-window id.
 constexpr UINT WM_EACP_TRAY = WM_APP + 0x10;
 constexpr UINT TRAY_ICON_ID = 1;
-
-namespace
-{
-// Builds a 32bpp ARGB HICON from the Image's straight RGBA pixels. The shell
-// scales it to the notification area's small-icon size, so a 16x16 or 32x32
-// source works best.
-HICON toHIcon(const Image& image)
-{
-    auto width = image.width();
-    auto height = image.height();
-
-    if (width <= 0 || height <= 0)
-        return nullptr;
-
-    BITMAPV5HEADER header = {};
-    header.bV5Size = sizeof(BITMAPV5HEADER);
-    header.bV5Width = width;
-    header.bV5Height = -height; // top-down
-    header.bV5Planes = 1;
-    header.bV5BitCount = 32;
-    header.bV5Compression = BI_BITFIELDS;
-    header.bV5RedMask = 0x00FF0000;
-    header.bV5GreenMask = 0x0000FF00;
-    header.bV5BlueMask = 0x000000FF;
-    header.bV5AlphaMask = 0xFF000000;
-
-    void* bits = nullptr;
-    auto hdc = GetDC(nullptr);
-    auto colorBitmap = CreateDIBSection(hdc,
-                                        reinterpret_cast<BITMAPINFO*>(&header),
-                                        DIB_RGB_COLORS,
-                                        &bits,
-                                        nullptr,
-                                        0);
-    ReleaseDC(nullptr, hdc);
-
-    if (!colorBitmap || !bits)
-    {
-        if (colorBitmap)
-            DeleteObject(colorBitmap);
-        return nullptr;
-    }
-
-    // RGBA (source) -> BGRA (DIB byte order).
-    auto* dst = static_cast<std::uint8_t*>(bits);
-    const auto* src = image.pixels().data();
-    eacp::simd::swapRedBlue(src, dst, (std::size_t) width * height);
-
-    auto maskBitmap = CreateBitmap(width, height, 1, 1, nullptr);
-
-    ICONINFO iconInfo = {};
-    iconInfo.fIcon = TRUE;
-    iconInfo.hbmColor = colorBitmap;
-    iconInfo.hbmMask = maskBitmap;
-
-    auto icon = CreateIconIndirect(&iconInfo);
-
-    DeleteObject(colorBitmap);
-    DeleteObject(maskBitmap);
-
-    return icon;
-}
-} // namespace
 
 struct TrayIcon::Native
 {
